+++ apps/web/public/locales/en/app.json
@@ -261,6 +261,7 @@
   "events": "Events",
   "eventStatusCanceled": "Canceled",
   "eventStatusPast": "Past",
+  "eventStatusUnconfirmed": "Unconfirmed",
   "eventStatusUpcoming": "Upcoming",
   "eventTypeActionsLabel": "Event type actions",
   "eventTypeDescriptionPlaceholder": "What is this event about?",
