--- apps/web/src/app/api/stripe/webhook/handlers/customer-subscription/created.ts
@@ -1,135 +1,0 @@
-import type { Stripe } from "@rallly/billing";
-import { prisma } from "@rallly/database";
-import { sendRawEmail } from "@rallly/emails";
-import { after } from "next/server";
-import { env } from "@/env";
-import { posthog } from "@/features/analytics/posthog";
-import { subscriptionMetadataSchema } from "@/features/subscription/schema";
-import {
-  getExpandedSubscription,
-  getSubscriptionDetails,
-  isSubscriptionActive,
-  syncSpaceTier,
-  toDate,
-} from "../utils";
-
-export async function onCustomerSubscriptionCreated(event: Stripe.Event) {
-  const subscription = await getExpandedSubscription(
-    (event.data.object as Stripe.Subscription).id,
-  );
-
-  const isActive = isSubscriptionActive(subscription);
-  const { priceId, currency, interval, amount } =
-    getSubscriptionDetails(subscription);
-
-  const res = subscriptionMetadataSchema.safeParse(subscription.metadata);
-
-  if (!res.success) {
-    throw new Error("Invalid subscription metadata");
-  }
-
-  const { userId, spaceId } = res.data;
-
-  const subscriptionItem = subscription.items.data[0];
-
-  if (!subscriptionItem) {
-    throw new Error(
-      `Missing subscription item in subscription ${subscription.id}`,
-    );
-  }
-
-  const subscriptionItemId = subscriptionItem.id;
-  const quantity = subscriptionItem.quantity ?? 1;
-
-  const tier = await prisma.$transaction(async (tx) => {
-    await tx.subscription.upsert({
-      where: { id: subscription.id },
-      create: {
-        id: subscription.id,
-        userId,
-        active: isActive,
-        quantity,
-        subscriptionItemId,
-        priceId,
-        currency,
-        interval,
-        amount,
-        status: subscription.status,
-        createdAt: toDate(subscription.created),
-        periodStart: toDate(subscription.current_period_start),
-        periodEnd: toDate(subscription.current_period_end),
-        cancelAtPeriodEnd: subscription.cancel_at_period_end,
-        spaceId,
-      },
-      update: {
-        id: subscription.id,
-        userId,
-        active: isActive,
-        subscriptionItemId,
-        quantity,
-        priceId,
-        currency,
-        interval,
-        amount,
-        status: subscription.status,
-        createdAt: toDate(subscription.created),
-        periodStart: toDate(subscription.current_period_start),
-        periodEnd: toDate(subscription.current_period_end),
-        cancelAtPeriodEnd: subscription.cancel_at_period_end,
-        spaceId,
-      },
-      include: {
-        space: true,
-      },
-    });
-
-    return syncSpaceTier(tx, spaceId);
-  });
-
-  posthog()?.groupIdentify({
-    groupType: "space",
-    groupKey: spaceId,
-    properties: {
-      seatCount: quantity,
-      tier,
-    },
-  });
-
-  posthog()?.capture({
-    distinctId: userId,
-    event: "subscription_create",
-    properties: {
-      interval,
-      $set: {
-        tier,
-      },
-    },
-    groups: {
-      space: spaceId,
-    },
-  });
-
-  if (isActive) {
-    after(async () => {
-      const user = await prisma.user.findUnique({
-        where: { id: userId },
-        select: { email: true, name: true },
-      });
-
-      if (user) {
-        await sendRawEmail({
-          to: user.email,
-          from: { name: "Luke from Rallly", address: env.SUPPORT_EMAIL },
-          subject: "Thanks for subscribing to Rallly Pro",
-          text: [
-            "Hey there,",
-            "Thanks so much for subscribing to Rallly Pro! I really appreciate your support.",
-            "I'd love to learn more about how you're using Rallly. What kind of events are you scheduling? Are there any features you wish Rallly had?",
-            "Just hit reply and let me know. I read every response and it helps shape what we build next.",
-            "Luke\nFounder, Rallly",
-          ].join("\n\n"),
-        });
-      }
-    });
-  }
-}
