--- apps/web/src/app/api/stripe/webhook/handlers/customer-subscription/deleted.ts
@@ -1,65 +1,0 @@
-import type { Stripe } from "@rallly/billing";
-import { stripe } from "@rallly/billing";
-import { prisma } from "@rallly/database";
-import { posthog } from "@/features/analytics/posthog";
-import { subscriptionMetadataSchema } from "@/features/subscription/schema";
-import { syncSpaceTier } from "../utils";
-
-export async function onCustomerSubscriptionDeleted(event: Stripe.Event) {
-  const subscription = await stripe.subscriptions.retrieve(
-    (event.data.object as Stripe.Subscription).id,
-  );
-
-  // void any unpaid invoices
-  const invoices = await stripe.invoices.list({
-    subscription: subscription.id,
-    status: "open",
-  });
-
-  for (const invoice of invoices.data) {
-    await stripe.invoices.voidInvoice(invoice.id);
-  }
-
-  const res = subscriptionMetadataSchema.safeParse(subscription.metadata);
-
-  if (!res.success) {
-    throw new Error("Invalid subscription metadata");
-  }
-
-  const { userId, spaceId } = res.data;
-
-  const tier = await prisma.$transaction(async (tx) => {
-    await tx.subscription.update({
-      where: {
-        id: subscription.id,
-      },
-      data: {
-        active: false,
-        status: "canceled",
-      },
-    });
-
-    return syncSpaceTier(tx, spaceId);
-  });
-
-  posthog()?.groupIdentify({
-    groupType: "space",
-    groupKey: spaceId,
-    properties: {
-      tier,
-    },
-  });
-
-  posthog()?.capture({
-    distinctId: userId,
-    event: "subscription_cancel",
-    properties: {
-      $set: {
-        tier,
-      },
-    },
-    groups: {
-      space: spaceId,
-    },
-  });
-}
