--- apps/web/src/app/api/stripe/webhook/handlers/customer/index.ts
@@ -1,2 +1,0 @@
-export * from "./created";
-export * from "./deleted";
