--- apps/web/src/app/api/stripe/webhook/handlers/payment-method/index.ts
@@ -1,3 +1,0 @@
-export * from "./attached";
-export * from "./detached";
-export * from "./updated";
